Handling of data with React Context and removed clutter
Data is being handled by the React Context component in order
to provide the entire frontend with access to the data, in lieu
of data drilling. Cell has been removed and replaced entirely with
the Item component, as it is more practical than to create a cell
for each item. Routing has also been slightly tweaked to allow the
navbar to persist throughout every page. Test data has also been added.